app/requirements.txt
charset-normalizer==3.3.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
Flask-Caching==2.3.0
pytest==8.3.2
pytest-flask==1.3.0